requirements.txt
groq
python-docx
sentence-transformers
numpy
PyPDF2
scikit-learn
flask
flask-cors
gunicorn
python-dotenv==1.0.1

